Rename artifacts for 3rd party dependencies according to major version.
