--- jodatime/src/main/java/org/jdbi/v3/jodatime/DateTimeMapper.java
@@ -1,30 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jdbi.v3.jodatime;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Timestamp;
-
-import org.jdbi.v3.core.statement.StatementContext;
-import org.jdbi.v3.core.mapper.ColumnMapper;
-import org.joda.time.DateTime;
-
-public class DateTimeMapper implements ColumnMapper<DateTime> {
-    @Override
-    public DateTime map(ResultSet r, int columnNumber, StatementContext ctx) throws SQLException {
-        Timestamp ts = r.getTimestamp(columnNumber);
-        return ts == null ? null : new DateTime(ts.getTime());
-    }
-}
