--- spring/src/test/java/org/jdbi/v3/spring/Callback.java
@@ -1,21 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jdbi.v3.spring;
-
-import org.jdbi.v3.core.Jdbi;
-
-public interface Callback
-{
-    void call(Jdbi dbi);
-}
