+++ spring4/src/test/java/org/jdbi/v3/spring4/Service.java
@@ -1,0 +1,30 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jdbi.v3.spring4;
+
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+
+public interface Service
+{
+    void inPropagationRequired(Callback c);
+
+    void inRequiresNew(Callback c);
+
+    void inNested(Callback c);
+
+    @Transactional(propagation= Propagation.REQUIRES_NEW, isolation = Isolation.READ_UNCOMMITTED)
+    void inRequiresNewReadUncommitted(Callback c);
+}
