+++ stringtemplate3/src/test/resources/org/jdbi/v3/stringtemplate3/TestStringTemplateGroupReference$Dao.sql.stg
